+++ src/timessquare/services/github/repo.py
@@ -41,6 +41,11 @@ from timessquare.domain.noteburst import NoteburstJobStatus
 from timessquare.domain.page import PageExecutionInfo, PageModel
 
 from ..page import PageService
+from .client import (
+    create_github_client,
+    create_github_installation_client,
+    get_app_jwt,
+)
 
 
 class GitHubRepoService:
@@ -72,6 +77,60 @@ class GitHubRepoService:
         self._github_client = github_client
         self._page_service = page_service
         self._logger = logger
+
+    @classmethod
+    async def create_for_repo(
+        cls,
+        *,
+        owner: str,
+        repo: str,
+        http_client: AsyncClient,
+        page_service: PageService,
+        logger: BoundLogger,
+    ) -> GitHubRepoService:
+        """Create a github repo service for a specific repository (requires
+        that the Times Square GitHub App is installed for that repository).
+
+        Parameters
+        ----------
+        owner : `str`
+            The GitHub repo's owner.
+        repo : `str`
+            The GitHub repo's name.
+        http_client : `AsyncClient`
+            An httpx client.
+        github_client : `GitHubAPI`
+            A GidgetHub API client that is authenticated as a GitHub app
+            installation.
+        page_service : `PageService`
+            The Page service. This GitHubRepoService acts as a layer on top of
+            the regular page service to handle domain models from the github
+            domain.
+        logger : `BoundLogger`
+            A logger, ideally with request/worker job context already bound.
+        """
+        app_jwt = get_app_jwt()
+        app_client = create_github_client(http_client=http_client)
+        installation_data = await app_client.getitem(
+            "/repos/{owner}/{repo}/installation",
+            url_vars={"owner": owner, "repo": repo},
+            jwt=app_jwt,
+        )
+        installation_id = installation_data["id"]
+        installation_client = await create_github_installation_client(
+            http_client=http_client, installation_id=installation_id
+        )
+        return cls(
+            http_client=http_client,
+            github_client=installation_client,
+            page_service=page_service,
+            logger=logger,
+        )
+
+    @property
+    def page_service(self) -> PageService:
+        """The page service used by the repo service."""
+        return self._page_service
 
     async def sync_from_repo_installation(
         self,
@@ -383,7 +442,7 @@ class GitHubRepoService:
         This check actually creates/updates Page resources, hence it is run
         at the service layer, rather than in a domain model.
         """
-        check = NotebookExecutionsCheck(check_run)
+        check = NotebookExecutionsCheck(check_run, repo)
         await check.submit_in_progress(self._github_client)
         self._logger.debug("Notebook executions check in progress")
 
@@ -505,3 +564,43 @@ class GitHubRepoService:
                 )
 
         await check.submit_conclusion(github_client=self._github_client)
+
+    async def get_check_runs(
+        self, owner: str, repo: str, head_sha: str
+    ) -> List[GitHubCheckRunModel]:
+        """Get the check runs from GitHub corresponding to a commit.
+
+        https://docs.github.com/en/rest/checks/runs#list-check-runs-for-a-git-reference
+        """
+        check_runs: List[GitHubCheckRunModel] = []
+        async for item in self._github_client.getiter(
+            "/repos/{owner}/{repo}/commits/{ref}/check-runs",
+            url_vars={"owner": owner, "repo": repo, "ref": head_sha},
+            iterable_key="check_runs",
+        ):
+            check_runs.append(GitHubCheckRunModel.parse_obj(item))
+        return check_runs
+
+    async def get_pulls_for_check_runs(
+        self, check_runs: List[GitHubCheckRunModel]
+    ) -> List[GitHubPullRequestModel]:
+        """Get the pull requests from GitHub covered by the provided check
+        runs.
+
+        Normally we'll look up the check runs first with `get_check_runs`
+        and then use this method to get information about the corresponding
+        pull requests.
+        """
+        # reduce the unique pull request urls
+        pr_urls: List[str] = []
+        for check_run in check_runs:
+            for pr in check_run.pull_requests:
+                pr_urls.append(pr.url)
+        pr_urls = sorted(list(set(pr_urls)))
+
+        pull_requests: List[GitHubPullRequestModel] = []
+        for pr_url in pr_urls:
+            data = await self._github_client.getitem(pr_url)
+            pull_requests.append(GitHubPullRequestModel.parse_obj(data))
+
+        return pull_requests